navbar creado y secciÃon results
