+++ package.json
@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@material-ui/core": "^4.12.4",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.26.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-redux": "^7.2.6",
     "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4",
-    "@material-ui/core": "4.12.4",
-    "axios": "^0.26.1",
     "redux": "^4.1.2",
     "redux-persist": "^6.0.0",
-    "react-redux": "^7.2.6",
-    "redux-thunk": "^2.4.1"
+    "redux-thunk": "^2.4.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
