+++ src/App.css
@@ -1,15 +1,8 @@
-.App {
-  text-align: center;
+.App{
+    background: #f3f3f3;
+    height: 100vh;
 }
 
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+.App h1{
+    font-weight: 400;
 }
